Refactor for equals
Some proposed changes to decrease cyclomatic complexity of equals. Mostly there is nothing to do to improve the function as the complexity comes from null-checks and other such precautions. The proposals are assuming some fundamental differences could be made to the rest of the code base, namely agreeing on a case convention for the project, now it needs to check for both snake and camelcase, also a huge improvement can be made if we assume all fields exist and don't have to do any null-checking. 

Closes #6